refactor: export API type contracts to shared package (#470)
Move 12 dependency-free API type files from server/types/api/ to shared/types/api/,
establishing shared/ as a proper TypeScript project (@peek/shared-types) that emits
.d.ts declarations. Server resolves shared types via tsconfig paths, with re-exporting
stubs preserving all existing controller imports. Eliminates entity type duplication
(server/types/entities.ts deleted) and removes the CI sync check script.
