+++ .gitignore
@@ -18,6 +18,7 @@ dist/
 build/
 */dist/
 */build/
+*.tsbuildinfo
 
 # Logs
 *.log
