+++ server/tsconfig.json
@@ -10,7 +10,11 @@
     "allowSyntheticDefaultImports": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
-    "noUncheckedIndexedAccess": true
+    "noUncheckedIndexedAccess": true,
+    "paths": {
+      "@peek/shared-types": ["../shared/dist/index.d.ts"],
+      "@peek/shared-types/*": ["../shared/dist/*"]
+    }
   },
   "include": ["**/*.ts"],
   "exclude": ["node_modules", "dist", "tests", "integration", "**/*.test.ts", "**/__tests__"]
