+++ shared/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@peek/shared-types",
+  "version": "0.0.1",
+  "private": true,
+  "type": "module",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "tsc"
+  },
+  "devDependencies": {
+    "typescript": "~5.8.3"
+  }
+}
